refactor the service implementation file using JpaRepository to manage data
