--- src/main/java/com/AmmarAli/Oferton/job/Job.java
@@ -1,69 +1,0 @@
-package com.AmmarAli.Oferton.job;
-
-import org.springframework.stereotype.Service;
-
-
-public class Job {
-    private Long id;
-    private String name;
-    private String description;
-    private String minSalary;
-    private  String maxSalary;
-    private String location;
-    public Job(Long id, String name, String description, String minSalary, String maxSalary, String location) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.minSalary = minSalary;
-        this.maxSalary = maxSalary;
-        this.location = location;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getMinSalary() {
-        return minSalary;
-    }
-
-    public void setMinSalary(String minSalary) {
-        this.minSalary = minSalary;
-    }
-
-    public String getMaxSalary() {
-        return maxSalary;
-    }
-
-    public void setMaxSalary(String maxSalary) {
-        this.maxSalary = maxSalary;
-    }
-
-    public String getLocation() {
-        return location;
-    }
-
-    public void setLocation(String location) {
-        this.location = location;
-    }
-}
