+++ src/main/java/com/AmmarAli/Oferton/job/repositories/JobService.java
@@ -1,5 +1,7 @@
-package com.AmmarAli.Oferton.job;
+package com.AmmarAli.Oferton.job.repositories;
 
+
+import com.AmmarAli.Oferton.job.entites.Job;
 
 import java.util.List;
 
@@ -7,9 +9,6 @@ public interface JobService {
     List<Job> findAll();
     void createJob(Job job);
     Job getJobsById(Long id);
-
     boolean deleteJobsById(Long id);
-
-
     boolean updateJob(Long id, Job updateJob);
 }
