+++ src/main/java/com/AmmarAli/Oferton/job/serviceImpl/JobServiceImpl.java
@@ -1,56 +1,53 @@
-package com.AmmarAli.Oferton.job.impl;
+package com.AmmarAli.Oferton.job.serviceImpl;
 
-import com.AmmarAli.Oferton.job.Job;
-import com.AmmarAli.Oferton.job.JobService;
+import com.AmmarAli.Oferton.job.entites.Job;
+import com.AmmarAli.Oferton.job.repositories.JobRepository;
+import com.AmmarAli.Oferton.job.repositories.JobService;
 import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class JobServiceImpl implements JobService {
-    private List<Job> jobs = new ArrayList<>();
+
+    //    private List<Job> jobs = new ArrayList<>();
     private Long nextId = 1L;
+    JobRepository jobRepository;
+    public JobServiceImpl(JobRepository jobRepository) {
+        this.jobRepository = jobRepository;
+    }
 
     @Override
     public List<Job> findAll() {
-        return jobs;
+        return jobRepository.findAll();
     }
 
     @Override
     public void createJob(Job job) {
         job.setId(nextId++);
-        jobs.add(job);
+        jobRepository.save(job);
     }
 
     @Override
     public Job getJobsById(Long id) {
-        for (Job job : jobs) {
-            if (job.getId().equals(id)) {
-                return job;
-            }
-        }
-        return null;
+        return  jobRepository.findById(id).orElse(null);
     }
 
     @Override
     public boolean deleteJobsById(Long id) {
-        Iterator<Job> iterator = jobs.iterator();
-        while (iterator.hasNext()){
-            Job job = iterator.next();
-            if (job.getId().equals(id)){
-                iterator.remove();
-                return true;
-            }
+        try {
+            jobRepository.deleteById(id);
+            return true;
+        }catch (Exception e){
+            return false;
         }
-        return false;
     }
 
     @Override
     public boolean updateJob(Long id, Job updateJob) {
-        for (Job job: jobs){
-            if (job.getId().equals(id)){
+        Optional<Job> jobOptional = jobRepository.findById(id);
+            if (jobOptional.isPresent()){
+                Job job = jobOptional.get();
                 job.setName(updateJob.getName());
                 job.setDescription(updateJob.getDescription());
                 job.setMinSalary(updateJob.getMinSalary());
@@ -58,7 +55,6 @@ public class JobServiceImpl implements JobService {
                 job.setLocation(updateJob.getLocation());
                 return true;
             }
-        }
         return false;
     }
